# optimize build document and array