@isopov 2025-07-09
Sorry, turns out these are not hot methods and probably not worth the complexity